SAMZA-2609: Data model changes to track lifecycle information of job coordinator (#1449)
Description:
With AM HA, the new AM will regenerate job model, do job planning and rewrite configurations and may have to spin up containers. The new rewritten configurations, job model can change from the existing metadata on which the containers are running. In order to handle this change of metadata, we need a way to identify lifecycle of the AM and store some information on the previous snapshot of the metadata.

Changes:
Epoch identifier of the AM attempt so that we can determine if the new attempt is part of the same lifecycle of AM or a new deployment lifecycle
Metadata on the job model, configuration so, that we can use this information to decide the course of action if current metadata changed in comparison with the previous snapshot.
Metrics capturing the changes in metadata
Coordinator stream message updates for new metadata
The code path as is not exercised actively and will be leveraged by the PRs to follow.

API Changes:
Added new Coordinator stream message type
Introduced data model for job coordinator metadata internal to samza