SAMZA-2610: Handle Metadata changes for AM HA orchestration (#1450)
Description:
AM performs planning and job model generation for every incarnation. With AM-HA, the new job model or configuration may invalidate the containers from the previous attempt. In order to ensure correctness, we handle this by detecting these changes and restart all the containers in case of any changes to metadata (job model or configuration).

Changes:

Detect changes in metadata by reading older metadata from coordinator stream and signal the CPM
As part of resource request & orchestration, ignore the containers that are already running from the previous attempt and proceed to release them if metadata changed.
Releasing the container will signal RM through AMRM client and RM will orchestrate killing the processing container. It is different from the normal StopStreamProcessor flow as the NMClient isn't the source of truth and doesn't have context about the containers spun in the previous attempts